make the server test generic
