+++ pkg/storage/unified/resource/server_test.go
@@ -241,285 +241,3 @@ func TestSimpleServer(t *testing.T) {
 		require.ErrorIs(t, err, ErrOptimisticLockingFailed)
 	})
 }
-
-func TestResourcePermissionScenarios(t *testing.T) {
-	// Test user
-	testUser := &identity.StaticRequester{
-		Type:           claims.TypeUser,
-		Login:          "testuser",
-		UserID:         123,
-		UserUID:        "u123",
-		OrgRole:        identity.RoleAdmin,
-		IsGrafanaAdmin: true,
-	}
-
-	testCases := []struct {
-		name           string
-		initialFolder  string
-		targetFolder   string
-		permissionMap  map[string]bool
-		expectSuccess  bool
-		expectedChecks int
-		expectedError  int32 // HTTP status code for error, 0 if no error expected
-	}{
-		{
-			name:           "Create resource in folder",
-			initialFolder:  "folder1",
-			targetFolder:   "", // No move, just create
-			permissionMap:  map[string]bool{"folder1:create": true},
-			expectSuccess:  true,
-			expectedChecks: 1,
-			expectedError:  0,
-		},
-		{
-			name:           "Create resource denied",
-			initialFolder:  "folder1",
-			targetFolder:   "", // No move, just create
-			permissionMap:  map[string]bool{"folder1:create": false},
-			expectSuccess:  false,
-			expectedChecks: 1,
-			expectedError:  http.StatusForbidden,
-		},
-		{
-			name:           "Update resource in same folder",
-			initialFolder:  "folder1",
-			targetFolder:   "folder1", // Same folder, just update
-			permissionMap:  map[string]bool{"folder1:update": true},
-			expectSuccess:  true,
-			expectedChecks: 1,
-			expectedError:  0,
-		},
-		{
-			name:           "Update resource denied",
-			initialFolder:  "folder1",
-			targetFolder:   "folder1", // Same folder
-			permissionMap:  map[string]bool{"folder1:update": false},
-			expectSuccess:  false,
-			expectedChecks: 1,
-			expectedError:  http.StatusForbidden,
-		},
-		{
-			name:           "Move resource to another folder - allowed",
-			initialFolder:  "folder1",
-			targetFolder:   "folder2", // Moving to a different folder
-			permissionMap:  map[string]bool{"folder1:update": true, "folder2:create": true},
-			expectSuccess:  true,
-			expectedChecks: 2, // Should check both folders
-			expectedError:  0,
-		},
-		{
-			name:           "Move resource - source folder access denied",
-			initialFolder:  "folder1",
-			targetFolder:   "folder2",
-			permissionMap:  map[string]bool{"folder1:update": false, "folder2:create": true},
-			expectSuccess:  false,
-			expectedChecks: 1, // Should stop at first check
-			expectedError:  http.StatusForbidden,
-		},
-		{
-			name:           "Move resource - destination folder access denied",
-			initialFolder:  "folder1",
-			targetFolder:   "folder2",
-			permissionMap:  map[string]bool{"folder1:update": true, "folder2:create": false},
-			expectSuccess:  false,
-			expectedChecks: 2, // Should do both checks but fail on second
-			expectedError:  http.StatusForbidden,
-		},
-		{
-			name:           "Move resource - from empty folder to named folder",
-			initialFolder:  "",
-			targetFolder:   "folder1",
-			permissionMap:  map[string]bool{":update": true, "folder1:create": true},
-			expectSuccess:  true,
-			expectedChecks: 2,
-			expectedError:  0,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Reset the backend for each test
-			store, err := NewCDKBackend(context.Background(), CDKBackendOptions{
-				Bucket: memblob.OpenBucket(nil),
-			})
-			require.NoError(t, err)
-
-			// Create a mock access client with the test case's permission map
-			checksPerformed := []claims.CheckRequest{}
-			mockAccess := &mockAccessClient{
-				allowed:    false, // Default to false
-				allowedMap: tc.permissionMap,
-				checkFn: func(req claims.CheckRequest) {
-					checksPerformed = append(checksPerformed, req)
-				},
-			}
-
-			server, err := NewResourceServer(ResourceServerOptions{
-				Backend:      store,
-				AccessClient: mockAccess,
-			})
-			require.NoError(t, err)
-
-			ctx := claims.WithAuthInfo(context.Background(), testUser)
-
-			// The resource key is the same for all tests
-			key := &ResourceKey{
-				Group:     "test.grafana.app",
-				Resource:  "testresources",
-				Namespace: "default",
-				Name:      "test-resource",
-			}
-
-			// If this is a create-only test
-			if tc.targetFolder == "" {
-				// Create resource
-				resourceJSON := fmt.Sprintf(`{
-					"apiVersion": "test.grafana.app/v1",
-					"kind": "TestResource",
-					"metadata": {
-						"name": "test-resource",
-						"uid": "test123",
-						"namespace": "default",
-						"annotations": {
-							"grafana.app/folder": "%s"
-						}
-					},
-					"spec": {
-						"title": "Test Resource"
-					}
-				}`, tc.initialFolder)
-
-				checksPerformed = []claims.CheckRequest{}
-				created, err := server.Create(ctx, &CreateRequest{
-					Value: []byte(resourceJSON),
-					Key:   key,
-				})
-				require.NoError(t, err)
-
-				if tc.expectSuccess {
-					require.Nil(t, created.Error)
-					require.True(t, created.ResourceVersion > 0)
-				} else {
-					require.NotNil(t, created.Error)
-					require.Equal(t, tc.expectedError, created.Error.Code)
-				}
-
-				require.Len(t, checksPerformed, tc.expectedChecks)
-				if len(checksPerformed) > 0 {
-					require.Equal(t, utils.VerbCreate, checksPerformed[0].Verb)
-					require.Equal(t, tc.initialFolder, checksPerformed[0].Folder)
-				}
-			} else {
-				// Create a resource first, then update/move it
-				initialResourceJSON := fmt.Sprintf(`{
-					"apiVersion": "test.grafana.app/v1",
-					"kind": "TestResource",
-					"metadata": {
-						"name": "test-resource",
-						"uid": "test123",
-						"namespace": "default",
-						"annotations": {
-							"grafana.app/folder": "%s"
-						}
-					},
-					"spec": {
-						"title": "Test Resource"
-					}
-				}`, tc.initialFolder)
-
-				// Override permissions for initial creation to always succeed
-				mockAccess.allowed = true
-				created, err := server.Create(ctx, &CreateRequest{
-					Value: []byte(initialResourceJSON),
-					Key:   key,
-				})
-				require.NoError(t, err)
-				require.Nil(t, created.Error)
-
-				// Now try the update/move with the configured permissions
-				targetResourceJSON := fmt.Sprintf(`{
-					"apiVersion": "test.grafana.app/v1",
-					"kind": "TestResource",
-					"metadata": {
-						"name": "test-resource",
-						"uid": "test123",
-						"namespace": "default",
-						"annotations": {
-							"grafana.app/folder": "%s"
-						}
-					},
-					"spec": {
-						"title": "Test Resource Updated"
-					}
-				}`, tc.targetFolder)
-
-				mockAccess.allowed = false // Reset to use the map
-				checksPerformed = []claims.CheckRequest{}
-
-				updated, err := server.Update(ctx, &UpdateRequest{
-					Key:             key,
-					Value:           []byte(targetResourceJSON),
-					ResourceVersion: created.ResourceVersion,
-				})
-				require.NoError(t, err)
-
-				if tc.expectSuccess {
-					require.Nil(t, updated.Error)
-					require.True(t, updated.ResourceVersion > created.ResourceVersion)
-				} else {
-					require.NotNil(t, updated.Error)
-					require.Equal(t, tc.expectedError, updated.Error.Code)
-				}
-
-				require.Len(t, checksPerformed, tc.expectedChecks)
-
-				// Verify the correct permission checks were made
-				if tc.initialFolder != tc.targetFolder && len(checksPerformed) >= 2 {
-					// This is a folder move operation
-					require.Equal(t, utils.VerbUpdate, checksPerformed[0].Verb)
-					require.Equal(t, tc.initialFolder, checksPerformed[0].Folder)
-
-					require.Equal(t, utils.VerbCreate, checksPerformed[1].Verb)
-					require.Equal(t, tc.targetFolder, checksPerformed[1].Folder)
-				} else if len(checksPerformed) > 0 {
-					// Regular update, no folder change
-					require.Equal(t, utils.VerbUpdate, checksPerformed[0].Verb)
-					require.Equal(t, tc.initialFolder, checksPerformed[0].Folder)
-				}
-			}
-		})
-	}
-}
-
-// Mock access client for testing
-type mockAccessClient struct {
-	allowed    bool
-	allowedMap map[string]bool
-	checkFn    func(claims.CheckRequest)
-}
-
-func (m *mockAccessClient) Check(ctx context.Context, user claims.AuthInfo, req claims.CheckRequest) (claims.CheckResponse, error) {
-	if m.checkFn != nil {
-		m.checkFn(req)
-	}
-
-	// Check specific folder:verb mappings if provided
-	if m.allowedMap != nil {
-		key := fmt.Sprintf("%s:%s", req.Folder, req.Verb)
-		if allowed, exists := m.allowedMap[key]; exists {
-			return claims.CheckResponse{Allowed: allowed}, nil
-		}
-	}
-
-	return claims.CheckResponse{Allowed: m.allowed}, nil
-}
-
-func (m *mockAccessClient) Compile(ctx context.Context, user claims.AuthInfo, req claims.ListRequest) (claims.ItemChecker, error) {
-	return func(name, folder string) bool {
-		key := fmt.Sprintf("%s:%s", folder, req.Verb)
-		if allowed, exists := m.allowedMap[key]; exists {
-			return allowed
-		}
-		return m.allowed
-	}, nil
-}
